+++ apps/web/src/app/product/[id]/api/index.ts
@@ -1,12 +1,10 @@
 import {
-  useQuery,
   useMutation,
   useSuspenseQuery,
   useSuspenseInfiniteQuery,
 } from '@tanstack/react-query';
 import { apiRequest } from '@/api/apiRequest';
 import { USER_QUERY_KEY } from '@/query-key/user';
-import { AUTH_QUERY_KEY } from '@/query-key/auth';
 import {
   ApiResponseBodyProductAvailableTimesResponseVoid,
   ApiResponseBodyProductClosedDatesResponseVoid,
@@ -14,7 +12,6 @@ import {
   ApiResponseBodyProductPeopleRangeResponseVoid,
   ApiResponseBodyProductReservationResponseVoid,
   ProductReservationRequest,
-  GetOnboardingData,
 } from '@/swagger-api';
 import { productDetailOptions, productPortfoliosOptions, productReviewsOptions } from './options';
 
@@ -127,35 +124,4 @@ export const useGetPortfolioList = (id: number, isLogIn: boolean) => {
 // 상품 리뷰 목록 조회 API
 export const useGetProductReviewList = (id: number) => {
   return useSuspenseInfiniteQuery(productReviewsOptions(id));
-};
-
-// 예약자 정보 조회 API
-export const useGetUsersOnboarding = (isLogIn: boolean) => {
-  return useQuery({
-    queryKey: AUTH_QUERY_KEY.ONBOARDING_USER(),
-    queryFn: async () => {
-      try {
-        const res = await apiRequest<GetOnboardingData>({
-          endPoint: '/api/v1/users/onboarding',
-          method: 'GET',
-        });
-
-        if (!res.data) {
-          throw new Error('/api/v1/users/onboarding 응답에 데이터가 존재하지 않습니다.');
-        }
-        return res.data;
-      } catch (error) {
-        if (typeof error === 'string') {
-          try {
-            const parsed = JSON.parse(error);
-            if (parsed.status === 404) return null;
-          } catch {
-            console.error(error);
-          }
-        }
-        throw error;
-      }
-    },
-    enabled: isLogIn,
-  });
 };
