+++ apps/web/src/auth/apis/index.ts
@@ -8,6 +8,7 @@ import {
   GetSwitchedUserProfileResponse,
   PatchUserRoleData,
   CreateKakaoLoginData,
+  GetOnboardingData
 } from '@/swagger-api';
 import { setAccessToken, getAccessToken, deleteAccessToken } from '../token';
 import { deleteAuthUser, setAuthUser } from '../userType';
@@ -136,5 +137,36 @@ export const useSwitchUserProfile = () => {
     onError: () => {
       error('프로필 전환에 실패했습니다. 잠시 후 다시 시도해주세요.', undefined, 'top-[2rem]');
     },
+  });
+};
+
+// 예약자 정보 조회 API
+export const useGetUsersOnboarding = (isLogIn: boolean) => {
+  return useQuery({
+    queryKey: AUTH_QUERY_KEY.ONBOARDING_USER(),
+    queryFn: async () => {
+      try {
+        const res = await apiRequest<GetOnboardingData>({
+          endPoint: '/api/v1/users/onboarding',
+          method: 'GET',
+        });
+
+        if (!res.data) {
+          throw new Error('/api/v1/users/onboarding 응답에 데이터가 존재하지 않습니다.');
+        }
+        return res.data;
+      } catch (error) {
+        if (typeof error === 'string') {
+          try {
+            const parsed = JSON.parse(error);
+            if (parsed.status === 404) return null;
+          } catch {
+            console.error(error);
+          }
+        }
+        throw error;
+      }
+    },
+    enabled: isLogIn,
   });
 };
